refactor: 💡 trade tab - trading input
