+++ src/constants/display.ts
@@ -3,3 +3,5 @@
 // };
 
 export const DELAYED_HIDDEN_SECONDS = 1000;
+
+export const TRADING_INPUT_STEP = 0.01;
